Eviction function changed (fixed a deadlock)
Under some circumstances, there will be a deadlock in eviction function
-- Suppose (1) there comes a query (A, B), and (2) the admission of
list(B) requires the eviction of list(A),  and (3) A and B belongs to
the same hash bucket, then the eviction of list(A) will not success
forever, because lock of slot(A) is locked already. In order to address
this problem, I allow the cache buffer to extend the MEMORYSIZE a little
(just in these particular circumstances). I don't think it will cause
any fatal harm, because it is just several MB (at most) more than the
designated cache size. However, I admitted, it is a little tricky here.